+++ cumulusci/core/tests/test_utils.py
@@ -5,6 +5,9 @@ import pytz
 
 from .. import utils
 
+from collections import OrderedDict
+from cumulusci.core.exceptions import ConfigMergeError
+
 
 class TestUtils(unittest.TestCase):
     def test_parse_datetime(self):
@@ -30,3 +33,29 @@ class TestUtils(unittest.TestCase):
         self.assertEqual(u"\xfc", utils.decode_to_unicode(b"\xfc"))
         self.assertEqual(u"\u2603", utils.decode_to_unicode(u"\u2603"))
         self.assertEqual(None, utils.decode_to_unicode(None))
+
+
+class TestMergedConfig(unittest.TestCase):
+    def test_init(self):
+        config = utils.merge_config(
+            OrderedDict(
+                [
+                    ("user_config", {"hello": "christian"}),
+                    ("global_config", {"hello": "world"}),
+                ]
+            )
+        )
+        self.assertEqual(config["hello"], "christian")
+
+    def test_merge_failure(self):
+        with self.assertRaises(ConfigMergeError) as cm:
+            config = utils.merge_config(
+                OrderedDict(
+                    [
+                        ("user_config", {"hello": "christian", "test": [1, 2]}),
+                        ("global_config", {"hello": "world", "test": {"sample": 1}}),
+                    ]
+                )
+            )
+        exception = cm.exception
+        self.assertEqual(exception.config_name, "user_config")
