+++ cumulusci/tasks/bulkdata.py
@@ -28,35 +28,13 @@ from sqlalchemy import event
 import requests
 import unicodecsv
 
-from cumulusci.core.utils import process_bool_arg
+from cumulusci.core.utils import process_bool_arg, ordered_yaml_load
 from cumulusci.core.exceptions import BulkDataException
 from cumulusci.tasks.salesforce import BaseSalesforceApiTask
 from cumulusci.utils import log_progress
 
 # TODO: UserID Catcher
 # TODO: Dater
-
-import yaml
-from collections import OrderedDict
-
-
-def ordered_load(stream, Loader=yaml.SafeLoader, object_pairs_hook=OrderedDict):
-    """ Load YAML file with OrderedDict, needed for Py2 
-    
-    code adapted from: https://stackoverflow.com/a/21912744/5042831
-    this could potentially be cleaned up to only run in py2 scenarios"""
-
-    class OrderedLoader(Loader):
-        pass
-
-    def construct_mapping(loader, node):
-        loader.flatten_mapping(node)
-        return object_pairs_hook(loader.construct_pairs(node))
-
-    OrderedLoader.add_constructor(
-        yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG, construct_mapping
-    )
-    return yaml.load(stream, OrderedLoader)
 
 
 # Create a custom sqlalchemy field type for sqlite datetime fields which are stored as integer of epoch time
@@ -512,7 +490,7 @@ class LoadData(BulkJobTaskMixin, BaseSalesforceApiTask):
 
     def _init_mapping(self):
         with open(self.options["mapping"], "r") as f:
-            self.mapping = ordered_load(f)
+            self.mapping = ordered_yaml_load(f)
 
 
 class QueryData(BulkJobTaskMixin, BaseSalesforceApiTask):
@@ -557,7 +535,7 @@ class QueryData(BulkJobTaskMixin, BaseSalesforceApiTask):
 
     def _init_mapping(self):
         with open(self.options["mapping"], "r") as f:
-            self.mappings = ordered_load(f)
+            self.mappings = ordered_yaml_load(f)
 
     def _soql_for_mapping(self, mapping):
         sf_object = mapping["sf_object"]
